refactor: move utility functions from control.rs to util.rs
